+++ NEWS.md
@@ -1,3 +1,9 @@
+# 2026-05-01
+
+For version `0.1.1`
+
+- In `wrap_caret`, feature matrix must be `x` instead of `X`
+
 # 2026-04-10
 
 - Initial version